Data-Engineering/1A-Cargar-Continente/requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
pyarrow